Start of refactor for parsing. Does not currently compile but am trying to move towards template declarations being more akin to how Rust and Java handles them (i.e. in line without an explicit template keyword), rather than the more verbose C++ syntax.
