+++ compiler/cx-compiler-ast/src/lib.rs
@@ -1,4 +1,9 @@
 #[macro_use]
-pub mod log;
+mod log;
+
+pub(crate) mod declarations;
+pub(crate) mod definitions;
+pub(crate) mod parsing_tools;
+
 pub mod parse;
 pub mod preparse;
