+++ compiler/cx-compiler-ast/src/parse.rs
@@ -1,15 +1,7 @@
-use cx_data_ast::parse::ast::CXAST;
-use cx_data_ast::parse::parser::ParserData;
+use cx_data_ast::parse::{ast::CXAST, parser::ParserData};
 use cx_data_lexer::TokenIter;
-use global_scope::parse_global_stmt;
 
-pub mod expression;
-pub mod global_scope;
-pub mod typing;
-pub mod operators;
-pub mod template;
-mod parsing_tools;
-mod structured_initialization;
+use crate::definitions::global_scope::parse_global_stmt;
 
 pub fn parse_ast(iter: TokenIter, base_ast: CXAST) -> Option<CXAST> {
     let mut data = ParserData {
@@ -27,4 +19,3 @@ pub fn parse_ast(iter: TokenIter, base_ast: CXAST) -> Option<CXAST> {
     
     Some(data.ast)
 }
-
