--- compiler/cx-compiler-ast/src/preparse/preparser.rs
@@ -1,358 +1,0 @@
-use cx_data_ast::{assert_token_matches, peek_next, try_next};
-use crate::preparse::typing::{parse_enum, parse_initializer, parse_params, parse_struct, parse_template_args, parse_union};
-use cx_data_lexer::token::{OperatorType, PunctuatorType, TokenKind};
-use cx_util::identifier::CXIdent;
-use cx_data_ast::parse::parser::{VisibilityMode};
-use cx_data_ast::preparse::CXNaiveFnIdent;
-use cx_data_ast::preparse::naive_types::{CXNaivePrototype, CXNaiveType, CXNaiveTypeKind, PredeclarationType};
-use cx_data_ast::preparse::templates::{CXFunctionTemplate, CXTemplatePrototype, CXTypeTemplate};
-use cx_data_lexer::{keyword, operator, punctuator, specifier, TokenIter};
-use cx_util::{log_error, CXResult};
-use crate::parse::global_scope::destructor_prototype;
-use crate::preparse::importing::parse_import;
-use crate::preparse::PreparseData;
-
-pub(crate) enum PreparseResult {
-    TypeDefinition(String, CXNaiveType),
-    FunctionDefinition(CXNaivePrototype),
-    DestructorDefinition(CXNaiveType),
-    TypeTemplate(CXTypeTemplate),
-    FunctionTemplate(CXFunctionTemplate),
-
-    Import(String),
-
-    Nothing
-}
-
-pub(crate) fn preparse_stmt(data: &mut PreparseData) -> Option<PreparseResult> {
-    match data.tokens.peek()?.kind {
-        keyword!(Template) => parse_template(data),
-        
-        keyword!(Import) => parse_import(&mut data.tokens),
-        keyword!(Struct, Enum, Union) => parse_plain_typedef(data),
-        keyword!(Typedef) => parse_typedef(&mut data.tokens),
-        
-        operator!(Tilda) => {
-            data.tokens.next();
-            let Some((None, _type)) = parse_initializer(&mut data.tokens) else {
-                log_preparse_error!(data.tokens, "Failed to parse type in destructor definition!");
-            };
-            goto_statement_end(&mut data.tokens);
-            
-            Some(PreparseResult::DestructorDefinition(_type))
-        },
-        
-        specifier!(Public) => {
-            data.tokens.next();
-            data.visibility_mode = VisibilityMode::Public;
-            try_next!(data.tokens, punctuator!(Colon));
-            
-            Some(PreparseResult::Nothing)
-        },
-        
-        specifier!(Private) => {
-            data.tokens.next();
-            data.visibility_mode = VisibilityMode::Private;
-            try_next!(data.tokens, punctuator!(Colon));
-            
-            Some(PreparseResult::Nothing)
-        },
-        
-        punctuator!(Semicolon) => {
-            data.tokens.next();
-            Some(PreparseResult::Nothing)
-        },
-
-        _ => preparse_global_expr(data),
-    }
-}
-
-pub(crate) fn parse_template(data: &mut PreparseData) -> Option<PreparseResult> {
-    assert_token_matches!(data.tokens, keyword!(Template));
-    assert_token_matches!(data.tokens, operator!(Less));
-    
-    let mut type_decls = Vec::new();
-    
-    loop {
-        assert_token_matches!(data.tokens, TokenKind::Identifier(template_name));
-        let template_name = template_name.clone();
-        
-        assert_token_matches!(data.tokens, punctuator!(Colon));
-        assert_token_matches!(data.tokens, keyword!(Type));
-    
-        type_decls.push(template_name);
-        
-        if !try_next!(data.tokens, operator!(Comma)) {
-            break;
-        }
-    }
-    
-    assert_token_matches!(data.tokens, operator!(Greater));
-    
-    let stmt = preparse_stmt(data)
-        .expect("Failed to parse global expression in template declaration!");
-    
-    match stmt {
-        PreparseResult::FunctionDefinition(signature) => {
-            Some(
-                PreparseResult::FunctionTemplate(
-                    CXFunctionTemplate {
-                        name: CXIdent::from(signature.name.mangle()),
-                        prototype: CXTemplatePrototype {
-                            types: type_decls.clone(),
-                        },
-                        shell: signature
-                    }
-                )
-            )
-        },
-
-        PreparseResult::TypeDefinition(name, type_) => {
-            Some(
-                PreparseResult::TypeTemplate(
-                    CXTypeTemplate {
-                        name: CXIdent::from(name),
-                        prototype: CXTemplatePrototype {
-                            types: type_decls.clone(),
-                        },
-                        shell: type_
-                    }
-                )
-            )
-        },
-
-        PreparseResult::DestructorDefinition(_type) => {
-            let prototype = destructor_prototype(_type.clone());
-
-            Some(
-                PreparseResult::FunctionTemplate(
-                    CXFunctionTemplate {
-                        name: CXIdent::from(prototype.name.mangle()),
-                        prototype: CXTemplatePrototype {
-                            types: type_decls.clone(),
-                        },
-                        shell: prototype
-                    }
-                )
-            )
-        },
-
-        _ => log_preparse_error!(data.tokens, "Invalid statement in template declaration!"),
-    }
-}
-
-pub(crate) fn parse_plain_typedef(data: &mut PreparseData) -> Option<PreparseResult> {
-    let starting_index = data.tokens.index;
-    
-    let (name, type_) = match &data.tokens.peek()?.kind {
-        keyword!(Struct) => parse_struct(&mut data.tokens)?,
-        keyword!(Enum) => parse_enum(&mut data.tokens)?,
-        keyword!(Union) => parse_union(&mut data.tokens)?,
-
-        tok => todo!("parse_plain_typedef: {tok:?}")
-    };
-
-    if matches!(type_, CXNaiveTypeKind::Identifier { predeclaration, .. } 
-        if predeclaration != PredeclarationType::None) {
-        data.tokens.index = starting_index; // Reset to the start if we didn't parse a type declaration
-        return preparse_global_expr(data);
-    }
-    
-    assert_token_matches!(data.tokens, punctuator!(Semicolon));
-    
-    Some(
-        PreparseResult::TypeDefinition(
-            name.unwrap().as_string(), 
-            type_.to_type()
-        )
-    )
-}
-
-pub(crate) fn parse_typedef(tokens: &mut TokenIter) -> Option<PreparseResult> {
-    assert_token_matches!(tokens, keyword!(Typedef));
-    let start_index = tokens.index;
-
-    let (name, type_) = parse_initializer(tokens)? else {
-        log_preparse_error!(tokens.with_index(start_index), "Could not parse typedef.");
-    };
-
-    let Some(name) = name else {
-        log_preparse_error!(tokens.with_index(start_index), "Typedef must have a name!");
-    };
-    
-    assert_token_matches!(tokens, punctuator!(Semicolon));
-    
-    Some(
-        PreparseResult::TypeDefinition (
-            name.to_string(),
-            type_
-        )
-    )
-}
-
-pub fn try_function_parse(tokens: &mut TokenIter, return_type: CXNaiveType, name: CXIdent) -> CXResult<Option<CXNaivePrototype>> {
-    match tokens.peek()?.kind {
-        // e.g:
-        // int main()
-        //         ^
-        TokenKind::Punctuator(PunctuatorType::OpenParen) => {
-            let Some(args) = parse_params(tokens) else {
-                log_preparse_error!(tokens, "Failed to parse parameters in function declaration!");
-            };
-            
-            Some(
-                Some(
-                    CXNaivePrototype {
-                        return_type,
-                        name: CXNaiveFnIdent::Standard(name),
-                        params: args.params,
-                        var_args: args.var_args,
-                        this_param: args.contains_this
-                    }
-                )
-            )
-        },
-
-        // e.g:
-        // void vec<int>::push()
-        //         ^
-        TokenKind::Operator(OperatorType::Less) => {
-            let Some(params) = parse_template_args(tokens) else {
-                return Some(None);
-            };
-            
-            let _type = name;
-
-            assert_token_matches!(tokens, TokenKind::Operator(OperatorType::ScopeRes));
-            assert_token_matches!(tokens, TokenKind::Identifier(name));
-            let fn_name = CXIdent::from(name.as_str());
-            
-            if !peek_next!(tokens, TokenKind::Punctuator(PunctuatorType::OpenParen)) {
-                log_preparse_error!(tokens, "Expected '(' after template arguments in function declaration!");
-            };
-
-            let Some(args) = parse_params(tokens) else {
-                log_preparse_error!(tokens, "Failed to parse parameters in function declaration!");
-            };
-            
-            let name = CXNaiveFnIdent::MemberFunction {
-                _type: CXNaiveTypeKind::TemplatedIdentifier {
-                    name: CXIdent::from(_type.as_str()),
-                    input: params
-                }.to_type(), 
-                function_name: fn_name
-            };
-
-            Some(
-                Some(
-                    CXNaivePrototype {
-                        return_type, name,
-                        params: args.params,
-                        var_args: args.var_args,
-                        this_param: args.contains_this
-                    }
-                )
-            )
-        }
-
-        // e.g:
-        // void renderer::draw()
-        //              ^
-        TokenKind::Operator(OperatorType::ScopeRes) => {
-            tokens.next();
-            let _type = name.as_string();
-            assert_token_matches!(tokens, TokenKind::Identifier(name));
-            let name = CXNaiveFnIdent::MemberFunction {
-                _type: CXNaiveTypeKind::Identifier {
-                    name: CXIdent::from(_type.as_str()),
-                    predeclaration: PredeclarationType::None
-                }.to_type(),
-                function_name: CXIdent::from(name.as_str())
-            };
-            let Some(params) = parse_params(tokens) else {
-                log_preparse_error!(tokens, "Failed to parse parameters in member function declaration!");
-            };
-
-            Some(
-                Some(
-                    CXNaivePrototype {
-                        return_type, name,
-                        params: params.params,
-                        var_args: params.var_args,
-                        this_param: params.contains_this
-                    }
-                )
-            )
-        },
-
-        _ => Some(None)
-    }
-}
-
-pub(crate) fn preparse_global_expr(data: &mut PreparseData) -> Option<PreparseResult> {
-    let (Some(name), return_type) = parse_initializer(&mut data.tokens)? else {
-        log_preparse_error!(data.tokens, "Invalid global expression, name not found!");
-    };
-    
-    if let Some(method) = try_function_parse(&mut data.tokens, return_type, name.clone())? {
-        goto_statement_end(&mut data.tokens);
-        
-        return Some(PreparseResult::FunctionDefinition(method))
-    };
-
-    // Global variables are parsed during the full parse, variable identifiers are not
-    // at least currently, needed in the process of full parsing an AST like type names are.
-
-    goto_statement_end(&mut data.tokens);
-    Some(PreparseResult::Nothing)
-}
-
-pub fn goto_statement_end(tokens: &mut TokenIter) -> Option<()> {
-    let mut bracket_stack = 0;
-    
-    while let Some(token) = tokens.next() {
-        match token.kind {
-            punctuator!(OpenBrace) => bracket_stack += 1,
-            punctuator!(CloseBrace) => {
-                bracket_stack -= 1;
-                
-                if bracket_stack == 0 {
-                    try_next!(tokens, punctuator!(Semicolon));
-                    break;
-                }
-            },
-            punctuator!(Semicolon) if bracket_stack == 0 => { break },
-            
-            _ => ()
-        }
-    }
-
-    Some(())
-}
-
-pub fn parse_intrinsic(tokens: &mut TokenIter) -> Option<CXIdent> {
-    let mut ss = String::new();
-
-    while let Some(TokenKind::Intrinsic(ident)) = tokens.peek().map(|tok| &tok.kind) {
-        ss.push_str(format!("{ident:?}").to_lowercase().as_str());
-        tokens.next();
-    }
-
-    if ss.is_empty() {
-        return None;
-    }
-
-    Some(
-        CXIdent::from(ss)
-    )
-}
-
-pub fn parse_std_ident(tokens: &mut TokenIter) -> Option<CXIdent> {
-    let TokenKind::Identifier(ident) = tokens.peek().cloned()?.kind else {
-        return None;
-    };
-
-    tokens.next();
-
-    Some(CXIdent::from(ident))
-}
